京东商城/requirements.txt
selenium==4.9.0
undetected-chromedriver==3.5.0
requests==2.28.2
fake-useragent==1.1.3
pillow==9.5.0
lxml==4.9.2
beautifulsoup4==4.12.2
scrapy==2.8.0
colorama==0.4.6
tqdm==4.65.0
pandas==2.0.1
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
python-dateutil==2.8.2
webdriver-manager==3.8.6
opencv-python==4.7.0.72
pyautogui==0.9.53
retry==0.9.2
urllib3==1.26.15
concurrent-log-handler==0.9.20 